moved spectral convetnions to type manager
