+++ dnora/type_manager/spectral_conventions.py
